Song Select + Editor Folder File Creation
The editor now asks and saves beatmap information, creates a folder in the beatmaps folder, creates difficulty files. These are then read in the song select screen which loads them, when the arrows keys are pressed then next song is loaded and difficulty, these can then be played. If a difficulty is not found they are disabled.