Online Signup/Login/Table/Leaderboard
Users can now sign up and login. When q is pressed in the editor a table is created from the song name and the username of the creator of the chart, this table name is saved in the beatmap file for reading to, which scores are uploaded after the user has finished playing in the results screen. Leaderboard data needs to be added and displayed.